Use wsl.exe instead of bash.exe for WSL command piping
bash.exe is deprecated according to the WSL documentation and wsl.exe what we're supposed to be using now.
An added benefit is the fact that we do not need to pass the -c option for running a single command
so we can drop the that and the associated quotations.